+++ src/banger_lib/fonts/__init__.py
@@ -9,12 +9,12 @@
 # @link      https://github.com/MarcinOrlowski/banger
 #
 ##################################################################################
+
+Font system for banger_lib.
 """
 
-# Import constants
 from .constants import DEFAULT_CHAR_SPACING, MANDATORY_CHARACTERS
 
-# Import core font system
 from .core import (
     CharacterData,
     FontMetadata,
@@ -24,14 +24,8 @@ from .core import (
     normalize_character_lines,
 )
 
-# Import fonts
-from .classic import ClassicFont
-from .matrix import MatrixFont
+from .factory import create_font, get_available_fonts, register_font
 
-# Import factory functions
-from .factory import create_font, get_available_fonts
-
-# Import API functions
 from .api import (
     get_font_height,
     get_font_characters,
@@ -40,37 +34,28 @@ from .api import (
     validate_font_character_coverage,
     get_all_fonts_validation_report,
     get_max_character_width,
+    _font_supports_lowercase,
+    _font_supports_uppercase,
 )
 
-"""Unified fonts package - clean and simple.
-
-This package contains all font-related functionality:
-- Core font system (interfaces, base classes, utilities)
-- Built-in font implementations
-- Font registry and management
-- Public API functions
-- Style-specific definitions
-"""
-
+from .ttf import TtfFont
 
 __all__ = [
     # Constants
     "DEFAULT_CHAR_SPACING",
     "MANDATORY_CHARACTERS",
-    # Core font system
+    # Core
     "CharacterData",
     "FontMetadata",
     "FontInterface",
     "BaseFont",
     "calculate_character_width",
     "normalize_character_lines",
-    # Built-in fonts
-    "ClassicFont",
-    "MatrixFont",
     # Factory
     "create_font",
     "get_available_fonts",
-    # API functions
+    "register_font",
+    # API
     "get_font_height",
     "get_font_characters",
     "get_character_data",
@@ -78,4 +63,8 @@ __all__ = [
     "validate_font_character_coverage",
     "get_all_fonts_validation_report",
     "get_max_character_width",
+    "_font_supports_lowercase",
+    "_font_supports_uppercase",
+    # TTF
+    "TtfFont",
 ]